requirements.txt
Django==4.2.23
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-decouple==3.8
django-allauth==0.58.2
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
psycopg2-binary==2.9.9
Pillow==11.2.1
stripe==12.2.0
asgiref==3.8.1
sqlparse==0.5.3
qrcode[pil]==8.2